Activity Log added (requries websockets)
Added an activity command that can be used to 'tail' the log output.
The basic command works and has formatted output but I have a lot to
learn about websockets and how to use this new module correctly in the
context of an otherwise blocking application.  Namely handling signals
is going to be tricky (ctrl+c).

Fixed up the media-refresh command a bit.